op42-move-spec: relocate alloc_for_spec_decode into mem_cache/allocation.py
